+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `events.Enter` and `events.Leave` events now bubble. https://github.com/Textualize/textual/pull/4818
 - Renamed `Widget.mouse_over` to `Widget.mouse_hover` https://github.com/Textualize/textual/pull/4818
 
+### Fixed
+
+- Fix crash when comparing `DataTable` `StringKey` objects https://github.com/Textualize/textual/pull/4820
+
 ## [0.74.0] - 2024-07-25
 
 ### Fixed